Breaking: Refactor DOM utils inside hint to `@hint/utils`
Fix #2192
