+++ packages/connector-jsdom/src/resource-loader.ts
@@ -3,10 +3,13 @@ import { URL } from 'url';
 import { debug as d } from '@hint/utils';
 import { ResourceLoader } from 'jsdom';
 
-import JSDOMConnector from './connector';
-import { FetchEnd, FetchError, HTMLDocument, NetworkData, utils } from 'hint';
+import { HTMLDocument, dom } from '@hint/utils';
+import { FetchEnd, FetchError, NetworkData, utils } from 'hint';
 
-const { dom: { createHTMLDocument, getElementByUrl }, contentType: { getContentTypeData, getType } } = utils;
+import JSDOMConnector from './connector';
+
+const { createHTMLDocument, getElementByUrl } = dom;
+const { contentType: { getContentTypeData, getType } } = utils;
 
 const debug: debug.IDebugger = d(__filename);
