+++ packages/hint-image-optimization-cloudinary/src/cloudinary-types.ts
@@ -1,5 +1,5 @@
 /* eslint-disable camelcase */
-import { HTMLElement } from 'hint';
+import { HTMLElement } from '@hint/utils/dist/src/dom/html';
 
 export type cloudinaryResult = {
     bytes: number;
