+++ packages/hint/package.json
@@ -28,8 +28,6 @@
     "mime-db": "1.35.0",
     "optionator": "^0.8.2",
     "ora": "^3.4.0",
-    "parse5": "^5.1.0",
-    "parse5-htmlparser2-tree-adapter": "^5.1.0",
     "semver": "^6.0.0",
     "update-notifier": "^2.5.0"
   },
@@ -40,8 +38,6 @@
     "@types/lodash": "^4.14.123",
     "@types/mime-db": "^1.27.0",
     "@types/node": "11.12.0",
-    "@types/parse5": "^5.0.0",
-    "@types/parse5-htmlparser2-tree-adapter": "^5.0.1",
     "@types/semver": "^6.0.0",
     "@types/sinon": "^7.0.11",
     "@types/update-notifier": "^2.5.0",
