+++ packages/hint/src/lib/types.ts
@@ -4,7 +4,6 @@ import { IParserConstructor } from './types/parser';
 import { IHintConstructor } from './types/hints';
 import { Severity } from './types/problems';
 
-export * from './types/html';
 export * from './types/connector';
 export * from './types/events';
 export * from './types/formatters';
