+++ packages/hint/src/lib/utils/index.ts
@@ -1,4 +1,3 @@
-import * as domUtils from './dom';
 import * as packagesUtils from './packages';
 import * as contentTypeUtils from './content-type';
 import * as resourceLoaderUtils from './resource-loader';
@@ -7,7 +6,6 @@ import * as jsonParserUtils from './json-parser';
 
 export * from './resource-loader';
 
-export const dom = domUtils;
 export const packages = packagesUtils;
 export const contentType = contentTypeUtils;
 export const resourceLoader = resourceLoaderUtils;
