+++ packages/parser-css/src/parser.ts
@@ -5,7 +5,8 @@ import * as postcss from 'postcss';
 // QUESTION: any way to use import { misc, logger } from '@hint/utils' without problems with webhint?
 import * as logger from '@hint/utils/dist/src/logging';
 import { normalizeString } from '@hint/utils/dist/src/misc/normalize-string';
-import { ElementFound, FetchEnd, Parser, HTMLElement } from 'hint/dist/src/lib/types';
+import { HTMLElement } from '@hint/utils/dist/src/dom/html';
+import { ElementFound, FetchEnd, Parser } from 'hint/dist/src/lib/types';
 import { StyleEvents } from './types';
 import { Engine } from 'hint';
