+++ packages/parser-html/src/parser.ts
@@ -26,9 +26,9 @@ try {
     // `canvas` is not installed, nothing to do
 }
 
+import { createHTMLDocument } from '@hint/utils/dist/src/dom/create-html-document';
 import { Parser, FetchEnd } from 'hint/dist/src/lib/types';
 import { Engine } from 'hint/dist/src/lib/engine';
-import { createHTMLDocument } from 'hint/dist/src/lib/utils/dom/create-html-document';
 import { HTMLEvents } from './types';
 
 export * from './types';
