+++ packages/utils/src/dom/index.ts
@@ -2,3 +2,4 @@ export * from './create-html-document';
 export * from './find-original-element';
 export * from './get-element-by-url';
 export * from './traverse';
+export * from './html';
