+++ packages/utils/src/index.ts
@@ -10,9 +10,13 @@ import * as networkUtils from './network';
 import * as npmUtils from './npm';
 import * as pkgs from './packages';
 import * as testUtils from './test';
+import * as domUtils from './dom';
 
 export * from './types/http-header';
 export * from './types/npm';
+export * from './types/html';
+export * from './types/problem-location';
+export * from './dom/html';
 
 export const packages = pkgs;
 export const fs = fsUtils;
@@ -24,6 +28,7 @@ export const appInsights = appInsightsUtils;
 export const caniuse = caniuseUtils;
 export const configStore = configStoreUtils;
 export const test = testUtils;
+export const dom = domUtils;
 
 export {
     asyncTry,
